requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Django==4.1.2
django-bootstrap4==22.2
django-filter==22.1
soupsieve==2.3.2.post1
sqlparse==0.4.3